requirements.txt
# Core Dependencies
pyspark==3.5.0
kafka-python==2.0.2
mlflow==2.9.2
pandas==2.1.4
numpy==1.26.2

# Graph Learning
torch==2.1.2
torch-geometric==2.5.0
dgl==1.1.3

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
pyarrow==14.0.1

# Monitoring & Logging
python-json-logger==2.0.7

